updated host to send data packet

+++ main/common/include/packet.h
@@ -46,7 +46,7 @@ typedef struct DataPacket_s
     double   altitude;
     uint64_t numTimeSamples;
     std::vector<double>* peak_timestamps;
-    std::vector<std::complex<short>>* waveformSamples;
+    std::vector<std::complex<float>>* waveformSamples;
 } DataPacket;
 
 #endif
